Collapse listeners to single class.
